+++ geotrek/sensitivity/filters.py
@@ -1,0 +1,10 @@
+from geotrek.common.filters import StructureRelatedFilterSet
+from .models import SensitiveArea
+
+
+class SensitiveAreaFilterSet(StructureRelatedFilterSet):
+    class Meta(StructureRelatedFilterSet.Meta):
+        model = SensitiveArea
+        fields = StructureRelatedFilterSet.Meta.fields + [
+            'species',
+        ]