+++ geotrek/sensitivity/forms.py
@@ -1,0 +1,10 @@
+from .models import SensitiveArea
+from geotrek.common.forms import CommonForm
+
+
+class SensitiveAreaForm(CommonForm):
+    geomfields = ['geom']
+
+    class Meta:
+        fields = ['species', 'geom']
+        model = SensitiveArea